+++ src/pages/Portal/index.js
@@ -14,10 +14,12 @@ import Widget from '../../molecules/Widget';
 import BoltArrowUpIcon from '../../atoms/BoltArrowUpIcon';
 import BoltArrowDownIcon from '../../atoms/BoltArrowDownIcon';
 import SunIcon from '../../atoms/SunIcon';
+import PowerPlugIcon from '../../atoms/PowerPlugIcon';
 
 import CurrentElectricityReceived from '../../molecules/CurrentElectricityReceived';
 import CurrentElectricityDelivered from '../../molecules/CurrentElectricityDelivered';
 import CurrentSolarPowerGenerating from '../../molecules/CurrentSolarPowerGenerating';
+import CurrentElectricityUsing from '../../molecules/CurrentElectricityUsing';
 
 export default function Portal() {
 	return (
@@ -35,6 +37,9 @@ export default function Portal() {
 				</Widget>
 				<Widget title="Solar generating" name="current-solar-generating" icon={<SunIcon />}>
 					<CurrentSolarPowerGenerating />
+				</Widget>
+				<Widget title="Using" name="current-electricity-using" icon={<PowerPlugIcon />}>
+					<CurrentElectricityUsing />
 				</Widget>
 			</WidgetGrid>
 		</LoggedOut>