+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.25.0] - 2022-01-11
+### Added
+- Electricity used today widget.
+- Electricity using current widget.
+
 ## [1.24.3] - 2022-01-09
 ### Fixed
 - Previous period for gas month chart, offset by one month.